requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
paho-mqtt>=1.6,<3.0
pydantic-settings>=2.2,<3.0
